// admin-ui/app/styles/types.ts
export type CardBorderStyleOptions = {
  isDark: boolean
  borderRadius?: number | string
  borderWidth?: string
  gradientPosition?: 'top left' | 'top right' | 'bottom left' | 'bottom right' | 'center'
  ellipseSize?: string
}

export type DynamicListColorOptions = {
  boxBg: string
  inputBg: string
  borderColor: string
  fontColor: string
  textMuted: string
  errorColor: string
}